feat: implement SnapCodeServer and refactor byte code retrieval in SnapServer
