+++ src/Nethermind/Nethermind.State.Flat/Sync/Snap/FlatSnapServer.cs
@@ -23,6 +23,7 @@ public class FlatSnapServer(
     public bool CanServe => true;
 
     private readonly ILogger _logger = logManager.GetClassLogger<FlatSnapServer>();
+    private readonly SnapCodeServer _codeServer = new(codeDb);
 
     private const long HardResponseByteLimit = 2000000;
     private const int HardResponseNodeLimit = 100000;
@@ -56,7 +57,8 @@ public class FlatSnapServer(
 
             byteLimit = Math.Max(Math.Min(byteLimit, HardResponseByteLimit), 1);
             int pathLength = pathSet.Count;
-            ArrayPoolList<byte[]> response = new(pathLength);
+            using DeferredRlpItemList.Builder builder = new(pathLength);
+            DeferredRlpItemList.Builder.Writer writer = builder.BeginRootContainer();
             ReadOnlyStateTrieStoreAdapter trieStore = new(bundle);
             StateTree tree = new(trieStore, logManager);
             bool abort = false;
@@ -68,13 +70,12 @@ public class FlatSnapServer(
                 switch (requestedPath.Length)
                 {
                     case 0:
-                        response.Dispose();
                         return null;
                     case 1:
                         try
                         {
                             byte[]? rlp = tree.GetNodeByPath(Nibbles.CompactToHexEncode(requestedPath[0]), stateId.StateRoot.ToCommitment());
-                            response.Add(rlp ?? []);
+                            writer.WriteValue(rlp);
                             responseSize += rlp?.Length ?? 0;
                         }
                         catch (MissingTrieNodeException)
@@ -98,7 +99,7 @@ public class FlatSnapServer(
                                 for (int reqStorage = 1; reqStorage < requestedPath.Length && responseSize < byteLimit && !cancellationToken.IsCancellationRequested; reqStorage++)
                                 {
                                     byte[]? sRlp = sTree.GetNodeByPath(Nibbles.CompactToHexEncode(requestedPath[reqStorage]));
-                                    response.Add(sRlp ?? []);
+                                    writer.WriteValue(sRlp);
                                     responseSize += sRlp?.Length ?? 0;
                                 }
                             }
@@ -111,44 +112,13 @@ public class FlatSnapServer(
                 }
             }
 
-            return response.Count == 0 ? EmptyByteArrayList.Instance : new ByteArrayListAdapter(response);
+            writer.Dispose();
+            return new RlpByteArrayList(builder.ToRlpItemList());
         }
     }
 
-    public IByteArrayList GetByteCodes(IReadOnlyList<ValueHash256> requestedHashes, long byteLimit, CancellationToken cancellationToken)
-    {
-        long currentByteCount = 0;
-        ArrayPoolList<byte[]> response = new(requestedHashes.Count);
-
-        if (byteLimit > HardResponseByteLimit)
-        {
-            byteLimit = HardResponseByteLimit;
-        }
-
-        foreach (ValueHash256 codeHash in requestedHashes)
-        {
-            if (currentByteCount > byteLimit || cancellationToken.IsCancellationRequested)
-            {
-                break;
-            }
-
-            if (codeHash.Bytes.SequenceEqual(Keccak.OfAnEmptyString.Bytes))
-            {
-                response.Add([]);
-                currentByteCount += 1;
-                continue;
-            }
-
-            byte[]? code = codeDb[codeHash.Bytes];
-            if (code is not null)
-            {
-                response.Add(code);
-                currentByteCount += code.Length;
-            }
-        }
-
-        return new ByteArrayListAdapter(response);
-    }
+    public IByteArrayList GetByteCodes(IReadOnlyList<ValueHash256> requestedHashes, long byteLimit, CancellationToken cancellationToken) =>
+        _codeServer.GetByteCodes(requestedHashes, byteLimit, cancellationToken);
 
     public (IOwnedReadOnlyList<PathWithAccount>, IByteArrayList) GetAccountRanges(
         Hash256 rootHash,
