+++ src/Nethermind/Nethermind.State/SnapServer/SnapServer.cs
@@ -36,7 +36,6 @@ public class SnapServer : ISnapServer
 
     private readonly IReadOnlyTrieStore _store;
     private readonly TrieStoreWithReadFlags _storeWithReadFlag;
-    private readonly IReadOnlyKeyValueStore _codeDb;
     private readonly ILogManager _logManager;
     private readonly ILogger _logger;
 
@@ -47,6 +46,7 @@ public class SnapServer : ISnapServer
 
     private readonly AccountDecoder _decoder = new();
     private readonly ILastNStateRootTracker? _lastNStateRootTracker;
+    private readonly SnapCodeServer _codeServer;
 
     private const long HardResponseByteLimit = 2000000;
     private const int HardResponseNodeLimit = 100000;
@@ -54,7 +54,7 @@ public class SnapServer : ISnapServer
     public SnapServer(IReadOnlyTrieStore trieStore, IReadOnlyKeyValueStore codeDb, ILogManager logManager, ILastNStateRootTracker? lastNStateRootTracker = null)
     {
         _store = trieStore ?? throw new ArgumentNullException(nameof(trieStore));
-        _codeDb = codeDb ?? throw new ArgumentNullException(nameof(codeDb));
+        _codeServer = new SnapCodeServer(codeDb ?? throw new ArgumentNullException(nameof(codeDb)));
         _lastNStateRootTracker = lastNStateRootTracker;
         _logManager = logManager ?? throw new ArgumentNullException(nameof(logManager));
         _logger = logManager.GetClassLogger<SnapServer>();
@@ -136,41 +136,8 @@ public class SnapServer : ISnapServer
         return new RlpByteArrayList(builder.ToRlpItemList());
     }
 
-    public IByteArrayList GetByteCodes(IReadOnlyList<ValueHash256> requestedHashes, long byteLimit, CancellationToken cancellationToken)
-    {
-        long currentByteCount = 0;
-        if (byteLimit > HardResponseByteLimit)
-        {
-            byteLimit = HardResponseByteLimit;
-        }
-
-        using DeferredRlpItemList.Builder builder = new(requestedHashes.Count);
-        DeferredRlpItemList.Builder.Writer writer = builder.BeginRootContainer();
-
-        foreach (ValueHash256 codeHash in requestedHashes)
-        {
-            // break when the response size exceeds the byteLimit - it is a soft limit
-            // so not a big issue if we so over slightly.
-            if (currentByteCount > byteLimit || cancellationToken.IsCancellationRequested) break;
-
-            if (codeHash.Bytes.SequenceEqual(Keccak.OfAnEmptyString.Bytes))
-            {
-                writer.WriteValue([]);
-                currentByteCount += 1;
-                continue;
-            }
-
-            byte[]? code = _codeDb[codeHash.Bytes];
-            if (code is not null)
-            {
-                writer.WriteValue(code);
-                currentByteCount += code.Length;
-            }
-        }
-
-        writer.Dispose();
-        return new RlpByteArrayList(builder.ToRlpItemList());
-    }
+    public IByteArrayList GetByteCodes(IReadOnlyList<ValueHash256> requestedHashes, long byteLimit, CancellationToken cancellationToken) =>
+        _codeServer.GetByteCodes(requestedHashes, byteLimit, cancellationToken);
 
     public (IOwnedReadOnlyList<PathWithAccount>, IByteArrayList) GetAccountRanges(
         Hash256 rootHash,
